submit

+++ srcs/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: macbook_air <macbook_air@student.42.fr>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/21 09:54:34 by macbook_air       #+#    #+#             */
-/*   Updated: 2021/10/21 09:54:35 by macbook_air      ###   ########.fr       */
+/*   Updated: 2021/11/26 06:19:13 by macbook_air      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int	count;
+	size_t	count;
 
 	count = 0;
 	while (str[count] != '\0')
-		count ++;
+		count++;
 	return (count);
 }
